Toxic content classification project/requirements.txt
streamlit
torch
transformers
pillow
pandas
scikit-learn
datasets
